modify api module
기존의 작성되어있던 API들을 모듈화하여 분리했습니다.
이에따라 기존에 사용되고 있던 경로들을 수정했습니다.
