+++ vue-diary-frontend/src/api/index.js
@@ -1,51 +1,22 @@
 import axios from 'axios';
 import { setInterceptors } from '@/api/common/interceptors';
 
+const createInstance = () => {
+  return axios.create({
+    baseURL: process.env.VUE_APP_API_URL,
+  });
+};
+
 /**
  * axios 설정 초기화 함수
  * @returns
  */
-const createInstance = () => {
+const createInstanceWithAuth = url => {
   return setInterceptors(
     axios.create({
-      baseURL: process.env.VUE_APP_API_URL,
+      baseURL: `${process.env.VUE_APP_API_URL}${url}`,
     }),
   );
 };
-const instance = createInstance();
-
-/**
- * 회원가입 API
- * @param {*} userData
- * @returns
- */
-const registerUser = userData => {
-  return instance.post('signup', userData);
-};
-
-/**
- * 로그인 API
- * @param {*} userData
- * @returns
- */
-const loginUser = userData => {
-  return instance.post('login', userData);
-};
-
-/**
- * 학습 노트 데이터를 조회하는 API
- */
-const fetchPosts = () => {
-  return instance.get('posts');
-};
-
-/**
- * 학습 노트 데이터를 생성하는 API
- * @param {*} postData
- * @returns
- */
-const createPost = postData => {
-  return instance.post('posts', postData);
-};
-
-export { registerUser, loginUser, fetchPosts, createPost };
+export const instance = createInstance();
+export const posts = createInstanceWithAuth('posts');
