+++ vue-diary-frontend/src/store/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 
-import { loginUser } from '@/api';
+import { loginUser } from '@/api/auth';
 import {
   getAuthFromCookie,
   getUserFromCookie,
