+++ vue-diary-frontend/src/utils/cookies.js
@@ -7,27 +7,15 @@ function saveUserToCookie(value) {
 }
 
 function getAuthFromCookie() {
-  return document.cookie.replace(
-    /(?:(?:^|.*;\s*)til_auth\s*=\s*([^;]*).*$)|^.*$/,
-    '$1',
-  );
+  return document.cookie.replace(/(?:(?:^|.*;\s*)til_auth\s*=\s*([^;]*).*$)|^.*$/, '$1');
 }
 
 function getUserFromCookie() {
-  return document.cookie.replace(
-    /(?:(?:^|.*;\s*)til_user\s*=\s*([^;]*).*$)|^.*$/,
-    '$1',
-  );
+  return document.cookie.replace(/(?:(?:^|.*;\s*)til_user\s*=\s*([^;]*).*$)|^.*$/, '$1');
 }
 
 function deleteCookie(value) {
   document.cookie = `${value}=; expires=Thu, 01 Jan 1970 00:00:01 GMT;`;
 }
 
-export {
-  saveAuthToCookie,
-  saveUserToCookie,
-  getAuthFromCookie,
-  getUserFromCookie,
-  deleteCookie,
-};
+export { saveAuthToCookie, saveUserToCookie, getAuthFromCookie, getUserFromCookie, deleteCookie };
